+++ templates/main.html
@@ -1,0 +1,10 @@
+{% if user.is_authenticated %}
+<h3>Login Success</h3>
+<p>Welcome!</p>
+<p><a href="{% url 'logout' %}">logout</a></p>
+{% else %}
+<p>
+  If you want to login, go to
+  <a href="{% url 'login' %}">here</a>
+</p>
+{% endif %}
