+++ templates/registration/logged_out.html
@@ -1,0 +1,5 @@
+<!--로그아웃 상태로 로그인 해야 접근 가능한 페이지 접근했을 경우 띄워주는 화면-->
+<div>
+  <p>You are not authenticated</p>
+  <p>Please login by clicking <a href="{% url 'login' %}">here</a></p>
+</div>
