+++ templates/registration/signup.html
@@ -1,0 +1,9 @@
+<h1>Register New User</h1>
+<form method="post">
+  {% csrf_token %} {{ form.as_p }}
+  <input type="submit" value="SignUp" />
+  <p>
+    If you already have an account,
+    <a href="{% url 'login' %}">login</a> instead
+  </p>
+</form>
